+++ todo-server/src/main/java/com/github/shirahata777/api/schedule/ScheduleService.java
@@ -15,8 +15,8 @@ import org.hibernate.cfg.Configuration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.github.shirahata777.dao.repository.SaveData;
 import com.github.shirahata777.dao.schedule.ScheduleTable;
-import com.github.shirahata777.repository.SaveData;
 
 import io.helidon.webserver.Routing;
 import io.helidon.webserver.ServerRequest;
