+++ todo-server/src/main/java/com/github/shirahata777/api/todo/TodoService.java
@@ -8,10 +8,10 @@ import javax.json.JsonObject;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.slf4j.LoggerFactory;
 
-import com.github.shirahata777.repository.schedule.SaveSchedule;
-import com.github.shirahata777.repository.todo.GetAllTodoData;
-import com.github.shirahata777.repository.todo.GetDetailTodoData;
-import com.github.shirahata777.repository.todo.SaveTodo;
+import com.github.shirahata777.dao.repository.schedule.SaveSchedule;
+import com.github.shirahata777.dao.repository.todo.GetAllTodoData;
+import com.github.shirahata777.dao.repository.todo.GetDetailTodoData;
+import com.github.shirahata777.dao.repository.todo.SaveTodo;
 
 import io.helidon.common.http.Parameters;
 import io.helidon.webserver.Routing;
