+++ todo-server/src/main/java/com/github/shirahata777/dao/repository/package-info.java
@@ -1,0 +1,1 @@
+package com.github.shirahata777.dao.repository;
