+++ todo-server/src/main/java/com/github/shirahata777/dao/repository/todo/SaveTodo.java
@@ -1,9 +1,9 @@
-package com.github.shirahata777.repository.todo;
+package com.github.shirahata777.dao.repository.todo;
 
 import javax.json.JsonObject;
 
+import com.github.shirahata777.dao.repository.SaveData;
 import com.github.shirahata777.dao.todo.TodoTable;
-import com.github.shirahata777.repository.SaveData;
 
 public class SaveTodo {
