--- todo-server/src/main/java/com/github/shirahata777/repository/package-info.java
@@ -1,1 +1,0 @@
-package com.github.shirahata777.repository;
